Bug changed update namaz

+++ apps/education/api/v1/serializers/namaz_serializer.py
@@ -24,7 +24,7 @@ class NamazReadSerializer(serializers.ModelSerializer):
 
 
 class NamazWriteSerializer(serializers.ModelSerializer):
-    images = ImageSerializer(many=True)
+    images = ImageSerializer(many=True, read_only=True)
 
     class Meta:
         model = Namaz
@@ -35,7 +35,7 @@ class NamazWriteSerializer(serializers.ModelSerializer):
         request = self.context.get("request")
         images_data = request.FILES.getlist("images", [])
 
-        if is_update:
+        if is_update and images_data:
             instance.images.all().delete()
 
         for image_data in images_data:
