+++ apps/service/api/v1/serializers/service_serializer.py
@@ -30,7 +30,7 @@ class ServiceReadSerializer(serializers.ModelSerializer):
 
 
 class ServiceWriteSerializer(serializers.ModelSerializer):
-    images = ImageSerializer(many=True)
+    images = ImageSerializer(many=True, read_only=True)
 
     class Meta:
         model = Service
@@ -41,7 +41,7 @@ class ServiceWriteSerializer(serializers.ModelSerializer):
         request = self.context.get("request")
         images_data = request.FILES.getlist("images", [])
 
-        if is_update:
+        if is_update and images_data:
             instance.images.all().delete()
 
         for image_data in images_data:
